+++ src/main/java/study/spring/seoulspring/controllers/HomeController.java
@@ -2,9 +2,13 @@ package study.spring.seoulspring.controllers;
 
 import java.text.DateFormat;
 import java.util.ArrayList;
+import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
 import java.util.Locale;
+import java.util.Map;
+
+import javax.servlet.http.HttpServletRequest;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -16,6 +20,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import study.spring.seoulspring.model.DateData;
 import study.spring.seoulspring.model.Department;
 import study.spring.seoulspring.model.View;
 import study.spring.seoulspring.service.DepartmentService;
@@ -115,6 +120,80 @@ public class HomeController {
 	@ResponseBody
 	public void ViewCountUpdate(String title) {
 		
+	}
+	
+	@RequestMapping(value = { "/reservation.do" }, method = RequestMethod.GET)
+	
+	public String reservation(Model model,@RequestParam("date") String date,@RequestParam("room_num") int room_num  ) {
+		
+		String month=date.substring(5,7);
+		String day =date.substring(8);
+		model.addAttribute("month", month);
+		model.addAttribute("day", day);
+		model.addAttribute("room_num", room_num);
+		
+		return "/reservation";
+		
+	}
+	
+	@RequestMapping(value = "/calendar.do", method = RequestMethod.GET)
+	public String calendar(Model model, HttpServletRequest request, DateData dateData,Locale locale){
+		
+		Calendar cal = Calendar.getInstance();
+		int today_year = cal.get(Calendar.YEAR);
+		int today_month = cal.get(Calendar.MONTH)+1;
+		String today ="";
+		int today_date = cal.get(Calendar.DATE);
+		if(today_month<10) {
+			 today = today_year +"-0"+today_month+"-"+today_date;
+		}
+		else {
+		 today = today_year +"-"+today_month+"-"+today_date;
+		}
+		DateData calendarData;
+		//검색 날짜
+		if(dateData.getDate().equals("")&&dateData.getMonth().equals("")){
+			dateData = new DateData(String.valueOf(cal.get(Calendar.YEAR)),String.valueOf(cal.get(Calendar.MONTH)),String.valueOf(cal.get(Calendar.DATE)),null);
+		}
+		//검색 날짜 end
+
+		Map<String, Integer> today_info =  dateData.today_info(dateData);
+		List<DateData> dateList = new ArrayList<DateData>();
+		
+		//실질적인 달력 데이터 리스트에 데이터 삽입 시작.
+		//일단 시작 인덱스까지 아무것도 없는 데이터 삽입
+		for(int i=1; i<today_info.get("start"); i++){
+			calendarData= new DateData(null, null, null, null);
+			dateList.add(calendarData);
+		}
+		
+		//날짜 삽입
+		for (int i = today_info.get("startDay"); i <= today_info.get("endDay"); i++) {
+			if(i==today_info.get("today")){
+				calendarData= new DateData(String.valueOf(dateData.getYear()), String.valueOf(dateData.getMonth()), String.valueOf(i), "today");
+			}else{
+				calendarData= new DateData(String.valueOf(dateData.getYear()), String.valueOf(dateData.getMonth()), String.valueOf(i), "normal_date");
+			}
+			dateList.add(calendarData);
+		}
+
+		//달력 빈곳 빈 데이터로 삽입
+		int index = 7-dateList.size()%7;
+		
+		if(dateList.size()%7!=0){
+			
+			for (int i = 0; i < index; i++) {
+				calendarData= new DateData(null, null, null, null);
+				dateList.add(calendarData);
+			}
+		}
+		System.out.println(dateList);
+		
+		//배열에 담음
+		model.addAttribute("dateList", dateList);		//날짜 데이터 배열
+		model.addAttribute("today_info", today_info);
+		model.addAttribute("today", today);
+		return "/calendar";
 	}
 
 }
