+++ src/main/java/study/spring/seoulspring/controllers/LoadmapController.java
@@ -68,53 +68,6 @@ public class LoadmapController {
 		return "loadmap/snu_class_loadmap_detail";
 	}
 	
-	@RequestMapping(value = "/calendar.do", method = RequestMethod.GET)
-	public String calendar(Model model, HttpServletRequest request, DateData dateData,Locale locale){
-		
-		Calendar cal = Calendar.getInstance();
-		DateData calendarData;
-		//검색 날짜
-		if(dateData.getDate().equals("")&&dateData.getMonth().equals("")){
-			dateData = new DateData(String.valueOf(cal.get(Calendar.YEAR)),String.valueOf(cal.get(Calendar.MONTH)),String.valueOf(cal.get(Calendar.DATE)),null);
-		}
-		//검색 날짜 end
-
-		Map<String, Integer> today_info =  dateData.today_info(dateData);
-		List<DateData> dateList = new ArrayList<DateData>();
-		
-		//실질적인 달력 데이터 리스트에 데이터 삽입 시작.
-		//일단 시작 인덱스까지 아무것도 없는 데이터 삽입
-		for(int i=1; i<today_info.get("start"); i++){
-			calendarData= new DateData(null, null, null, null);
-			dateList.add(calendarData);
-		}
-		
-		//날짜 삽입
-		for (int i = today_info.get("startDay"); i <= today_info.get("endDay"); i++) {
-			if(i==today_info.get("today")){
-				calendarData= new DateData(String.valueOf(dateData.getYear()), String.valueOf(dateData.getMonth()), String.valueOf(i), "today");
-			}else{
-				calendarData= new DateData(String.valueOf(dateData.getYear()), String.valueOf(dateData.getMonth()), String.valueOf(i), "normal_date");
-			}
-			dateList.add(calendarData);
-		}
-
-		//달력 빈곳 빈 데이터로 삽입
-		int index = 7-dateList.size()%7;
-		
-		if(dateList.size()%7!=0){
-			
-			for (int i = 0; i < index; i++) {
-				calendarData= new DateData(null, null, null, null);
-				dateList.add(calendarData);
-			}
-		}
-		System.out.println(dateList);
-		
-		//배열에 담음
-		model.addAttribute("dateList", dateList);		//날짜 데이터 배열
-		model.addAttribute("today_info", today_info);
-		return "/calendar";
-	}
+	
 
 }
