+++ src/main/webapp/WEB-INF/views/assets/css/style.css
@@ -18,7 +18,9 @@ body {
     display: none;
   }
 }
-
+:root{
+  --main-color:#0f0f70;
+}
 @media (max-width: 900px) {
   .snu_mobile_box {
     width: 100%;
